added cross_fit_type and variance_type argument in policy_eval
cross_fit_type: only relevant for target = "sub_effect". Should the estimate be calculated for each fold and then be average, or should the doubly robust score be pooled first and estimate then calculated for the pooled score.

variance_type: Should the IC be stacked, should the doubly robust score be pooled first, or should a new model be fitted to the complete data in order to calculate the IC.